requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
httpx>=0.27.0
websockets>=13.0
python-multipart>=0.0.9
pyyaml>=6.0
